requirements.txt
numpy>=1.24
h5py>=3.8
matplotlib>=3.7
pytest>=7.0
pandas>=2.0
pyarrow>=14.0
PyYAML>=6.0
